6/7 Continued work on the solution to prevent gravitational grind
It is technically straightforward to inhibit the effect of gravity between two (at_rest & touching) particles. In velocity verlet method, one can inhibit for the affected pair that gravitational forces are added to the velocity vectors.

The simulation of scenario 1 shows that particles almost immediately come to a halt and stay there. This could either be the known gravitational grind, or something more recently wrong.

However, I find that regardless of doing this, the velocity verlet method also updates the positions of the particles. Problem here is that step 4 seems to start with a situation where there is 0 relative velocity in the normal direction (towards the normal line), and after adding the velocities to positions, ends up with a distance >2 and a positive relative velocity in the normal direction. is the total relative velocity changed? Some reflection needed.

Small victory throughout: there were occasions where relative velocities were calculated inaccurately/wrongly, e.g. confusing the idea of relative velocities (which is a vector) with relative speed (which is a scalar/magnitude). The logic to inhibit collissions (e.g. fail the collission check if particles are (at_rest & touching) ) is now consistent.